backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
openai>=1.0.0
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-genai
python-dotenv>=1.0.0
